requirements.txt
python-telegram-bot>=21.0
groq>=0.11.0
python-dotenv==1.0.0
